II interface v2
Strawman PR to discuss a possible v2 II interface.
The interface introduces the following changes:
* remove any mention of anchor. The new API is based around "identity".
  (I decided against "user", as a user could have many identities.)
* remove the concept of "device" (the new API is based around keys)
* by doing the above two renames, all methods now have a new name
  that does not clash with the v1 api (except for `create_challenge`
  which has been renamed to `create_captcha`). This avoids having to
  introduce a `_v2` suffix (or similar).
* all changed methods now return `opt variant`, which allows to
  properly return errors (and also extend the types later without
  breaking clients)
* `update` is now clearly scoped to only allow changing metadata
* the `last_usage` timestamp can now be provided on write. The
  write will be rejected if the timestamp does not match the canister
  side value. This can be used to prevent writes of stale data.
* Some redundant type aliases have been removed.
* No changes have been made to the endpoints related to delegations
  as well as the non user-facing the technical ones
  (i.e. stats and archive related endpoints).

The Plan is to gradually introduce the new API alongside the old one
in an experimental state. Once we are happy with the new interface,
we can stabilize the API and deprecate (and later remove) the current
one.

This work is motivated by repeatedly bumping into limitations with the
current API, the latest two examples being the inability to introduce
proper errors and the need to introduce additional WebAuthnSignatures
(which causes problems on the current `update` call as it mixes sensitive
and non-sensitive data changes).

This PR is _not_ intended to be merged, but rather to form agreement on
the new API, so that we can then introduce it gradually as needed in
separate PRs.