Center CAPTCHA on page resize
On mobile devices, opening the virtual keyboard tends to push the
CAPTCHA out of view. This ensures the CAPTCHA is recentered when the
page is resized (e.g. when the virtual keyboard opens).

This is implemented via a window resize handler, which deregisters
itself as soon as it can.

This also works on non-mobile although should never really kick-in.