Show custom landing page for known dapps (#1590)
* Show custom landing page for known dapps

This adds a few variations to the landing page, esp. to show some
prettier information (logo, actual name instead of just URL) for known
dapps (i.e. dapps we use for the explorer).

This also adds a new screenshot for every page that contains a chasm,
that shows the chasm open.

This includes a workaround where the frontend uses the
`II_DUMMY_CAPTCHA` value to assume the dapp is in a "test" environment,
in which case it includes an extra "known dapp" for testing purposes.

* Fixup