Don't duplicate BASE_URL in dapp logo
This fixes an issue where the BASE_URL is prepended twice to the dapp
logo URL. The BASE_URL is already included in the `dapps.ts` module,
which loads the dapps.

This issue turns `/foo.png` into `//foo.png`, which browsers then interpret as
`https://foo.png` instead of `https://<this website>/foo.png`.