Add authn_method_add to v2 API
This PR adds the method `authn_method_add` to the II canister.
This required to place some restrictions on the `metadata` entries
that are mapped back to the `Device` type. However, once we migrate
the stable memory to a type that more closely reflects the API v2,
we should be able to lift this restriction.