Improve mobile screenshots
This increases the pixel density for mobile screenshots, which had
become not-so-useful because of their size. Even though technically the
specified device was an "iPhone SE" in practice it did not represent the
dimensions of an "iPhone SE". The new dimensions specified here make the
screenshots more representative of what II looks like on small modern
devices like the "iPhone SE".