Skip rust toolchain on test archive cache hit
This PR optimizes the CI pipeline to not compile
any tests at all, if the rust code has not changed.
Instead, the compiled tests from the last commit that
changed rust code will be reused.

This paves the way to simplify the rust caching and hopefully
get rid of mtime fiddling again.