Use test binary for canister sig checking (#1649)
* Use test binary for canister sig checking

The test-state-machine-binary has recently been extended
to offer canister signature checks. This allows us to remove
all remaining dependencies on the IC repository.

In particular, this removes the duplicate package warnings from
our build and drops a lot of dependencies (-> faster builds).

* Add JS rand support

* Add custom get_random for wasm

This is now explicit because previously we had a transitive
dependency on ic-crypto-getrandom-for-wasm which dealt with that
for us.

* Implement review feedback

* Use doc link to spec instead of code

* Implement review comments