Add scheduled RC releases
This deploys the latest `production` build on `main` to one of our test
canisters, every Tuesday evening (Zurich time).

This works by fetching all `production` build artifacts on the `main`
branch and grabbing the latest one; then it reuses the infrastructure we
have in place to deploy to mainnet test canisters.

Finally, a message is sent to our private slack.