Support multiple char replacements
This adds support for mapping multiple chars to one char in the CAPTCHA
(e.g. `1`, `I`, `l` etc can all map to `i`).

This is done by reworking the `CHAR_REPLACEMENTS` map to read the
"replacement" chars from its keys and "replaced" chars from its values
(now an array).

Additionally, this extracts the font creation/parsing to a (lazy) static
value, so that the font doesn't have to be recreated every time.