Extract identity number from word list
This splits the identity number input from the grid of recovery words.
Having the identity number as part of the grid was confusing for some
users.

The treatment is applied on three different screens (all screens that
show the word grid):

* When the recovery phrase is created, the identity is separate (though
  still copied to the clipboard)
* When confirming/verifying the recovery phrase, the identity is
  separate and uses colors for validity feedback (unlike words that also
  use an icon). We may revisit this for more clarity.
* When recovering using the phrase, the identity input is also split
  out.

The `anchorInput` component is adapted to accomodate e.g. more classes
and more callbacks (the new screens now make use of `oninput`).

The `Chan` gets a new helper, `zip()`, to zip two `Chan`s together.