Focus identity field on recovery
Previously, the identity number was just another word in the recoery
word grid, so in order to focus it we just focused the first word
(during recovery & new phrase confirmation).

Now that the identity is a different field, it is that new field that
should be focused; not the first BIP39 word.