Don't commit screenshots to repository
This removes the GHA job that commits screenshots to the branch, and
introduces a new workflow, running on PRs, that instead updates the PR's
description to show which screens have changed.

This prevents the repository from getting bloated further with
screenshots, and allows us to add more "screenshots" to track visual
changes better.