Read user number from recovery phrase input (#1720)
This changes the recovery phrase screens to not require the userNumber
to be known before the user types their phrase. Instead, the number is
read from the input.

This is part of the ongoing work for simplifying the recovery flow. By
not requiring the user number to be known, we'll be able to skip the
user number prompt step.