Continue with another device copy changes
Copy changes for the happy path when pressing the
"continue with another device" button on the landing page.

Includes a change to always prompt the anchor pickor with
additional text.

The error message has not been changed, since it is used in
multiple places and would first need to be split.