Make the browser storage key key type writeable
A new key type has been introduced internally (for rollback reasons)
in the last release. Now that it has been rolled out, the external
API is extended to allow reading and writing that key type.
It is intended to be used for the new browser storage based
authentication mechanism.

This PR introduces a breaking change on the candid interface:
* Old clients reading the new key type variant will fail when
  decoding the II responses. However, since the variant has to
  be written before it can be read, this should not cause issues.
* The test detecting breaking changes has been disabled until the
  next II release.
* The new API (v2) is designed in a way that interface extensions,
  such as this one, will no longer cause breaking changes.