Add support for dfx pull
This PR adds support for DFX pull. Supporting
dfx pull creates a dependency between the dev build
and the prod build, because the wasm metadata needs to include
the dev build hash.

Furthermore, it is not only the prod_build but also the dev build
that needs to include additional metadata.

To make the feature work this PR:
- changes the build order of the wasm modules so that the prod buid
  is built after the dev build to include the dev build hash
- all II wasm files include additional metadata, _if_ the commit
  corresponds to a release (since no valid metadata can be generated
  for non-released wasms)
- a script has been added to generate the above metadata
- a script for wasm hash verification has been added since verification
  of wasm hashes is no more difficult (as it entails building II twice
  and including the hash of one module in the other)

Future work:
- Refactor build scripts to separate building the wasm from adding metadata.
  This would allow to build all wasms in parallel again, shortening CI times.
- Try to move to `dfx` for wasm post-processing. This would remove a
  substantial amount of scripts / logic from our build process.