Introduce helper macro for handling variant results
This PR introduces a macro `cast!` which allows casting
an enum to a specific value (checked at runtime). This
is useful especially for handling API v2 call results as they
are all `opt varinat`.

The due to the way our formatter behaves, this macro does not actually
lead to less lines, but much shorter ones.