Break down c-input class
This breaks down the big `c-input` class into more specific parts.

In particular, the fullwidth aspect of `c-input` is extracted into a
separate class to that `c-input` can also be used with non-full-width
elements.

All attributes pertaining to the recovery words are moved to a new
class, `c-output--recovery` (which really is not an input, hence the
name change).