Add better PIN component for device verification (#1813)
This introduces a new components for PIN input (`pinInput`) and updates
`verifyTentativeDevice` to use this new component.

The component supports both typing & pasting the PIN. The PIN will be
autosubmitting when all inputs are filled, at most once. On subsequent
attempts the user will need to submit manually.

See tests & notes in code for the many corner cases & tweaks.