Fix outdated outputs on release build check step (#1819)
I forgot to update the `outputs` section of the `latest-release`
release build check CI workflow. This is the fix for that.

Additionally, there was a mixup between the dev and prod wasms,
so I updated the naming to make it more clear as well.

However, the workflow should not rely on the dev build at all.
This issue will be resolved in a separate PR.