Switch to cargo feature resolver 2
This PR switches the workspace to the new feature resolver v2.
It makes cleverer choices about the features to be included with
different builds.
In practice this reduces the size of the prod wasm by ~20kb and
speeds up the build by a few seconds.
See more information here: https://blog.rust-lang.org/2021/03/25/Rust-1.51.0.html#cargos-new-feature-resolver