Clean up close icon (#1825)
We use a "close" (X) icon in toasts and in the dapps explorer.

The previous cross was pretty much an upper case X, which looked a bit
odd. The scaling & the animating was inconsistent.

This replaces the base icon with a proper cross and makes the
proportions & animations a bit nicer.