Add more toast variations (#1832)
This restructures the toast styling implementation and adds some toast
variations.

In particular, the "irregularity" classes are renamed to "toast-body" to
be more accurate, and slightly modified to accomodate more toast
variations.

Said variations are added (for "info" and "success" toasts), and toasts
are now added to the showcase components page.

The design of "info" and "success" is not pretty, but these are
currently only used in the showcase.