Showcase registration flow
This refactors the `register` function to be able to showcase the
registration flow without needing to spin up a canister.

A type parameter had to be added to `RegisterResult` to allow "mocking"
the result without creating an actual canister connection.

The flow part is extracted from `register`, which is now a wrapper
around this new `registerFlow`; `register` now only deals with
canister/concrete implementation details whereas `registerFlow` only
focuses on the user interaction.