Rework authentication flows (#1847)
Reword authentication flows

This cleans up the authentication flows, and clarifies some of the
errors.

This mostly deals with mocking as many operations (like registering an
identity) as possible to make the `authenticateBoxFlow` as powerful &
mockable as possible.

This also fixes a long-standing issue where canceling the recovery would
just reload the page.