Add screens for non-passkey flow (#1862)
This introduces all the screens necessary for the non-passkey
(non-webauthn) flow. It includes screens for creating & using a PIN, as
well as changes to the CAPTCHA to make the stepper generic (as the
passkey & non-passkey flows both use the captcha but use different
steppers).

The CAPTCHA is also removed from the flows, since it's not a flow per
se.