Highlight passkey warning only on 0 passkeys
This changes the warning highlight around passkeys to only
appear if the number of passkeys is zero. Also, there is now a
text shown for passkeys in the non-warning state.

Together with #1888 this shifts the warning box highlight on an
identity with just a single passkey from the passkey box to the
recovery box.