Remove recovery phrase warning banner
The recovery phrase warning banner is no longer justified as the
domain migration has been put on hold. In addition, a bug (the banner
being added twice) has been reported.

The simplest solution to address this is to simply remove the banner.
There are other nag screens / warnings still in place.