Do not clear localstorage on logout (#1929)
* Do not clear localstorage on logout

Deleting identity numbers on logout is unexpected and
destructive. A change was requested to no longer do that.

Instead, we should give users a different way of managing the
numbers shown on the landing page (to be solved as a separate PR).

Note: 3 lines were changed in the production code. The rest is adapting
the e2e tests to use the pick identity flow instead of entering it new.

* Rename changed flows