Simplify canister ID injection & SRI
This updates the transforming that we do of HTML files. This
transforming handles injecting the canister ID & figuring out the
correct SRI directive.

Instead of having a custom `#setupJs` element in `index.html` that we
used to replace, we now directly use the `<script>` tag generated by
vitejs during the build. This means we do not need to strip the vitejs
script tag anymore (i.e. no need for the strip plugin).

Moreover, the `src` of the canister-injected tag is fixed to point to
the same absolute path as vite (was `index.js`, is now `/index.js`).