Add issuer lookup (#1994)
This adds an issuer lookup to the II webapp. The lookup works
by querying the issuer origin, expected to be the boundary node, and
reads the resulting `x-ic-canister-id` response header. This canister ID
is then used to perform the canister calls necessary for VC.

This also updates the VC test app to add an issuer field, and adds an
issuer build & deployment to the canister (selenium) tests.