Port components to Astro (#1997)
* Port components to Astro

This ports all components shown in the showcase to Astro components.

This means we now have proper routing for displaying those components.

The components are wrapped in native/web components so that Astro can
import them properly; the alternative of creating an arbitrary div with
`id="pageContent"` like we do elsewhere is problematic because Astro
bundles all javascript scripts together when e.g. using `Astro.glob`, meaning all scripts would
trigger on any page importing the components (like the index) and all
would fight for the same `#pageContent`.

The approach of creating new elements is a bit more robust, though also
a bit more verbose.

Another benefit of this approach is that we can e.g. embed/showcase all
components in the index page.

* Restructure showcase components