Split storage ops from serialization
This rearranges the storage (was: `userNumber`) module to improve
maintainability.

In particular, the localStorage-specific operations are extracted into
dedicated localStorage functions, and the actual
pruning/migrating/updating is also split out. This gives three layers:
localStorage, pruning/migrating/updating, and top-level functions (like
`getAnchors`) used by the rest of the codebase.

The storage (was: `userNumber`) test suite is also updated to clarify
that the test data is related to localStorage.