Migrate localStorage data to IndexedDB (#2040)
* Migrate localStorage data to IndexedDB

This ports our localStorage data (identity numbers used & timestamps) to
IndexedDB. By doing this we follow the team's convention of storing only
UI related, short-lived data in localStorage, and all meaningful data in
IndexedDB. This also simplifies the storage handling by skipping the
JSON serialization.

For the time being, the data is still written to localStorage in case a
rollback in needed. That way, data written after the upgrade but before
the rollback will still be available to users after the rollback.

* Use idb.clear()

* Use explicit idb-keyval imports