Bump vitest
This bumps vitest to work around a bug in recent Node: https://github.com/vitest-dev/vitest/issues/3826

The bump caused some failures in the crypto ed25519 tests on node where
subtle did not like an ArrayBuffer as an argument. It turns out that it
does not need a buffer at all and the original array can be supplied
instead.

Finally the order of test commands in the workflow is changed to fail
early if the fast tests fail.