Simplify canister ID lookup
This simplifies the way canister IDs are looked up, both for static
config & dynamically in the devserver.

In general, canister IDs are not looked up from the canister_ids.json
anymore, but instead `dfx canister id <CANISTER_NAME>` is used (so that
we don't rely on dfx internal).

The dynamic replica-forwarding is changed to avoid having to specify
canister IDs; instead domains can be specified which are treated as
`<canister ID>.<domain>` and forwarded to the appropriate canister.