Deploy issuer & test app canisters on release (#2101)
This adds automated deploys of the issuer & of the test app upon new II
releases. The issuer and test app can then be used to test the latest II
changes.

NOTE: the provisioning script for the issuer was updated to allow
running against a mainnet deploy.
NOTE: the test app & issuer canisters don't (yet?) have logic to point
to e.g. the test II on mainnet (the URL has to be set manually).