Update vc_issuer and test-app to use asset_util
This PR refactors both the vc_issuer as well as the test-app to make
use of the asset_util crate.

The asset_util crate is extended to allow updating assets after the
initial certification.

Includes a minor refactoring of asset_util which makes the
`add_certification_vX` part of the `CertifiedAssets` impl.

This PR is step **4** of 4:
1. move code from `assets.rs` to the crate
2. move certification header code from `http.rs` to the crate
3. refactor the crate a bit to make the public interface nicer
4. use the crate in the other canisters